CIVISM/ml/requirements.txt
# CIVISIM ML/AI Requirements
# Python 3.8+

# Base scientific computing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep learning (CPU-optimized)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# NLP & Text Processing
transformers>=4.30.0
pdfplumber>=0.9.0
python-docx>=0.8.11
spacy>=3.5.0

# Data visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
python-dotenv>=1.0.0
requests>=2.28.0
tqdm>=4.65.0
